+++ spec/helpers/country_helper_spec.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+require "rails_helper"
+
+RSpec.describe Country, type: :helper do
+  it "shengen area should contain only defined countries" do
+    schengen_countries = Country.countries_for_region("Schengen Area")
+    expect(schengen_countries.any?(&:nil?)).to be_falsey
+  end
+end